update the suggestions and the corddialog tool
