--- what3words/what3words/extlibs/qgiscommons-2.0.12.dist-info/RECORD
@@ -1,38 +1,0 @@
-qgiscommons-2.0.12.dist-info/INSTALLER,sha256=zuuue4knoyJ-UwPPXg8fezS7VCrXJQrAP7zeNuwvFQg,4
-qgiscommons-2.0.12.dist-info/METADATA,sha256=QKbIuiSxVpaLigchQbK0n-x_pMouWKd-JdJqfp4O68s,297
-qgiscommons-2.0.12.dist-info/RECORD,,
-qgiscommons-2.0.12.dist-info/WHEEL,sha256=v8slff5hmCpvciQ3G55d2d1CnOBupjDFJHDE2dUb1Ao,97
-qgiscommons-2.0.12.dist-info/top_level.txt,sha256=QesJKuGrLYUy_JWmMeKEpwvGzvMWMG_CAoXjt5awjIM,13
-qgiscommons2/__init__.py,sha256=bqPJXgkuNn23M-KSveRDzeOFJA7mODPncCLJZwIguOw,22
-qgiscommons2/__pycache__/__init__.cpython-38.pyc,,
-qgiscommons2/__pycache__/files.cpython-38.pyc,,
-qgiscommons2/__pycache__/layers.cpython-38.pyc,,
-qgiscommons2/__pycache__/settings.cpython-38.pyc,,
-qgiscommons2/__pycache__/utils.cpython-38.pyc,,
-qgiscommons2/files.py,sha256=pTWr_17jjE9JRYdQoLk3bmkZ9SE9fVXczLLzfR3P5xQ,1477
-qgiscommons2/gui/__init__.py,sha256=-nLYpwDcxuRReJPs8FRdf6oTPWjiXE_P_PUbyxG96lU,10431
-qgiscommons2/gui/__pycache__/__init__.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/authconfigselect.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/checkcombobox.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/iconlineedit.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/paramdialog.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/passwordlineedit.cpython-38.pyc,,
-qgiscommons2/gui/__pycache__/settings.cpython-38.pyc,,
-qgiscommons2/gui/authconfigselect.py,sha256=lekyn0eS8vUyo14tCQqfy4avrescQdwENzKBbvQxhNs,3864
-qgiscommons2/gui/checkcombobox.py,sha256=FARn7kgxJOLgR3rmlbJ19SPiXXP7lW5vT37trs0ECQo,5266
-qgiscommons2/gui/iconlineedit.py,sha256=qirQm9YxMkiC1htmzlO8mHt2XNS7kWUHAUKjYgc9Sr0,2195
-qgiscommons2/gui/paramdialog.py,sha256=z6EWr2Fk4ZpFd2BsyZU3jbUD7MtwLViVQdMeK6-EaZs,7711
-qgiscommons2/gui/passwordlineedit.py,sha256=P-wIn9Tfbcc8fBk78boKF7qhj_3zYNb1DnhkBUF7lH8,3496
-qgiscommons2/gui/settings.py,sha256=TbN-gfg1k91C7SRnywFmkP7AOBOqdf8i5giRLgw-QYA,14193
-qgiscommons2/icons/eye-slash.svg,sha256=g_xX4IXRF5fw7I8ERZtsxoO7IwZg-P97MXUoO4L2jT4,1600
-qgiscommons2/icons/eye.svg,sha256=vAMoMcT7R_nUbwqyM-9iMQ-QE53UurB3JtjyZ1uk-14,1159
-qgiscommons2/icons/lock.svg,sha256=Lk3rWs6EY_sC-B-nEiLuQjxEw9R3Zh1IXedrPJsLWak,667
-qgiscommons2/layers.py,sha256=ouap2M5YWwgRg29GVQfnkTTZk9sUGhl3JExKmUfDn-s,9074
-qgiscommons2/network/__init__.py,sha256=47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU,0
-qgiscommons2/network/__pycache__/__init__.cpython-38.pyc,,
-qgiscommons2/network/__pycache__/networkaccessmanager.cpython-38.pyc,,
-qgiscommons2/network/__pycache__/oauth2.cpython-38.pyc,,
-qgiscommons2/network/networkaccessmanager.py,sha256=ANoAtLtUKUEeL8BxYNTB0CcY7qb2OjObxkDfFPrczlQ,15277
-qgiscommons2/network/oauth2.py,sha256=0b9YXRxzkKEfvEoz_R0w5ZCg6RG4L7xnwq6v--onpAM,3516
-qgiscommons2/settings.py,sha256=aIUshbm83T1CvELuC-0LrdWt17O02hEU2WnK0xTRebs,4816
-qgiscommons2/utils.py,sha256=xkJWF2EU1EJrJiyrf7rhVzbWDKaVw-TpeePZCQrYxJY,3169
