--- what3words/what3words/extlibs/qgiscommons-2.0.12.dist-info/WHEEL
@@ -1,5 +1,0 @@
-Wheel-Version: 1.0
-Generator: bdist_wheel (0.33.6)
-Root-Is-Purelib: true
-Tag: py3-none-any
-
