--- what3words/what3words/extlibs/qgiscommons-2.0.12.dist-info/top_level.txt
@@ -1,1 +1,0 @@
-qgiscommons2
