--- what3words/what3words/extlibs/qgiscommons2/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "2.0.12"
