--- what3words/what3words/processingprovider/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# (c) 2016 Boundless, http://boundlessgeo.com
-# This code is licensed under the GPL 2.0 license.
-#
