--- what3words/what3words/settings.json
@@ -1,16 +1,0 @@
-[
-    {"name":"apiKey",
-     "label": "API Key",
-     "description": "what3words API key",
-     "type": "string",
-     "default": "",
-     "group": "General"
-    },
-	{"name":"addressLanguage",
-     "label": "what3words address language (2 Digit language code)",
-     "description": "what3words address language (2 Digit language code)",
-     "type": "string",
-     "default": "EN",
-     "group": "General"
-    }
-]
